api/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
pgvector==0.2.5

# Redis
redis==5.0.1

# OpenAI
openai==1.12.0
tiktoken==0.5.2

# PDF Processing
pymupdf==1.23.22
pdfplumber==0.10.4

# Utils
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
tenacity==8.2.3

# Dev
pytest==8.0.0
pytest-asyncio==0.23.4
